+++ lib/imprint.rb
@@ -4,6 +4,7 @@ require 'imprint/version'
 require 'imprint/signer'
 require 'imprint/renderer'
 require 'imprint/engine' if defined?(Rails)
+require 'imprint/railtie' if defined?(Rails)
 
 module Imprint
   def self.sign(source:, watermark:, expires_in:)