+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-## [Unreleased]
+## Changelog.md
 
+
+## 0.1.1
+- Rails integration documentation
+- Optional Railtie for Rails environments
 ## [0.1.0] - 2026-01-20
-
-- Initial release